+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added Grid tabs using Grid Filters (raoulvdberge)
 - The Grid now resizes based on screen size (max rows can be configured) (raoulvdberge)
 - Added configuration option to enable large fonts in Grid (raoulvdberge)
+- Made Solderer tooltip less big (raoulvdberge)
 
 ### 1.3.3
 - Updated Forge to 2188 (raoulvdberge)